+++ libcaf_core/caf/detail/parser/chars.hpp
@@ -35,6 +35,8 @@ inline bool in_whitelist(bool (*filter)(char), char ch) noexcept {
   return filter(ch);
 }
 
+CAF_CORE_EXPORT extern const char whitespace_chars[7];
+
 CAF_CORE_EXPORT extern const char alphanumeric_chars[63];
 
 CAF_CORE_EXPORT extern const char alphabetic_chars[53];